better renamming and restructuring of components
