+++ src/components/landing/ClockIcon.vue
